requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
cffi==1.17.1
cryptography==44.0.2
linkify-it-py==2.0.3
markdown-it-py==3.0.0
mdit-py-plugins==0.4.2
mdurl==0.1.2
platformdirs==4.3.7
pycparser==2.22
Pygments==2.19.1
pyperclip==1.9.0
rich==14.0.0
textual==3.0.1
typing_extensions==4.13.1
uc-micro-py==1.0.3
